+++ src/test/java/ChangeUserTests.java
@@ -32,15 +32,4 @@ public class ChangeUserTests extends BaseTest {
             response.then().statusCode(401);
             response.then().body("message", equalTo("You should be authorised"));
     }
-
-    @After
-    public void cleanUp() {
-
-        Response response = userClient.checkLoginUserApi(user);
-        bearerToken = response.then().extract().jsonPath().getString("accessToken");
-
-        if (bearerToken != null) {
-            userClient.checkDeleteUserWithAuthApi(bearerToken);
-        }
-    }
 }
