+++ src/test/java/LoginTests.java
@@ -26,17 +26,6 @@ public class LoginTests extends BaseTest {
         response.then().statusCode(401);
         response.then().body("message", equalTo("email or password are incorrect"));
     }
-
-    @After
-    public void cleanUp() {
-
-        Response response = userClient.checkLoginUserApi(user);
-        bearerToken = response.then().extract().jsonPath().getString("accessToken");
-
-        if (bearerToken != null) {
-            userClient.checkDeleteUserWithAuthApi(bearerToken);
-        }
-    }
 }
